feat(dashboard): extract Vite client
